+++ CHANGELOG.md
@@ -58,7 +58,8 @@
 
   ([Andrey Kozhev](https://github.com/ankddev))
 
-- When attempting to publish a package on Hex with an already taken name, the massage is clearer.
+- When attempting to publish a package on Hex with an already taken name,
+  the massage is clearer.
   ([vyacheslavhere](https://github.com/vyacheslavhere))
 
 ### Language server